+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 📆 v0.2.0 (03/12/2025)
+## 📆 v0.2.1 (03/13/2025)
+
+- Introduced the `update_todays_note` tool, which allows updating the contents of the daily note
+- Introduced the `delete_daily_note` tool, which allows deleting a specific daily note by date
+
+## 📆 v0.2.0 (03/13/2025)
 
 - MCP clients will now be notified whenever a gist is added/delete/duplicated/renamed. That way the resources list can be properly re-fetched.